Trunk: merged branch native-wrapper.
git-svn-id: svn://localhost/public/MoSync/trunk@1758 6ad7a17e-d552-4a1e-ace4-f958fe14b117

+++ runtimes/cpp/platforms/symbian/src/SyscallImpl.cpp
@@ -1248,7 +1248,7 @@ SYSCALL(int, maInvokeExtension(int, int, int, int)) {
 	BIG_PHAT_ERROR(ERR_FUNCTION_UNIMPLEMENTED);
 }
 
-SYSCALL(int, maIOCtl(int function, int a, int b, int c)) {
+SYSCALL(longlong, maIOCtl(int function, int a, int b, int c)) {
 	//move to individual functions?
 	if(gBtState != eAvailable) switch(function) {
 	case maIOCtl_maBtStartDeviceDiscovery:
